Add more client auth 1.1 unit tests and fix null filter director bug and origin service returns 501 bug.
