+++ project-set/components/client-auth/src/test/java/com/rackspace/papi/components/clientauth/rackspace/v1_1/RackspaceAuthenticationHandlerTest.java
@@ -5,107 +5,304 @@ import com.rackspace.auth.v1_1.AuthenticationServiceClient;
 import com.rackspace.papi.commons.util.http.CommonHttpHeader;
 import com.rackspace.papi.commons.util.http.HttpStatusCode;
 import com.rackspace.papi.commons.util.http.PowerApiHeader;
+import com.rackspace.papi.commons.util.servlet.http.ReadableHttpServletResponse;
 import com.rackspace.papi.components.clientauth.rackspace.config.AccountMapping;
 import com.rackspace.papi.components.clientauth.rackspace.config.AccountType;
+import com.rackspace.papi.components.clientauth.rackspace.config.AuthenticationServer;
 import com.rackspace.papi.components.clientauth.rackspace.config.RackspaceAuth;
 import com.rackspace.papi.filter.logic.FilterAction;
 import com.rackspace.papi.filter.logic.FilterDirector;
 import com.rackspacecloud.docs.auth.api.v1.Group;
 import com.rackspacecloud.docs.auth.api.v1.GroupsList;
-import java.util.Map;
-import java.util.Set;
-import javax.servlet.http.HttpServletRequest;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
 
+import javax.servlet.http.HttpServletRequest;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyString;
+import static org.mockito.Mockito.mock;
+
 import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 
 /**
- *
- * @author zinic
+ * @author fran
  */
 @RunWith(Enclosed.class)
 public class RackspaceAuthenticationHandlerTest {
 
-    public static class WhenAuthenticating {
+    @Ignore
+    public static abstract class TestParent {
 
-        private AuthenticationServiceClient authServiceClient;
-        private HttpServletRequest request;
-        private RackspaceAuthenticationHandler authHandler;
-        private GroupsList groups;
+        protected HttpServletRequest request;
+        protected ReadableHttpServletResponse response;
+        protected AuthenticationServiceClient authServiceClient;
+        protected RackspaceAuthenticationHandler handler;
+        protected RackspaceAuth rackAuthConfig;
+        protected GroupsList groups;
 
         @Before
-        public void standUp() {
-            final RackspaceAuth authConfig = new RackspaceAuth();
-
-            // Add a new account mapping to the account mapping list
-            final AccountMapping accountUriMapping = new AccountMapping();
-            accountUriMapping.setIdRegex(".*/([\\d]+)/.*");
-            accountUriMapping.setType(AccountType.MOSSO);
-
-            authConfig.getAccountMapping().add(accountUriMapping);
+        public void beforeAny() {
+            request = mock(HttpServletRequest.class);
+            response = mock(ReadableHttpServletResponse.class);
 
             // Add a default group to the groups list
             groups = new GroupsList();
-            
+
             final Group defaultGroup = new Group();
             defaultGroup.setDescription("A default group");
             defaultGroup.setId("group-id");
-            
+
             groups.getGroup().add(defaultGroup);
-            
-            // Mocking fun
+
+            // Setup config
+            rackAuthConfig = new RackspaceAuth();
+            rackAuthConfig.setDelegatable(delegatable());
+
+            final AccountMapping mapping = new AccountMapping();
+            mapping.setIdRegex("/start/(.*)/");
+            mapping.setType(AccountType.CLOUD);
+
+            rackAuthConfig.getAccountMapping().add(mapping);
+
+            final AuthenticationServer authenticationServer = new AuthenticationServer();
+            authenticationServer.setUri("http://some.auth.endpoint");
+            rackAuthConfig.setAuthenticationServer(authenticationServer);
+
             authServiceClient = mock(AuthenticationServiceClient.class);
-            authHandler = new RackspaceAuthenticationHandler(authConfig, authServiceClient);
+            handler = new RackspaceAuthenticationHandler(rackAuthConfig, authServiceClient);
+        }
 
-            // Mocking out the request
-            request = mock(HttpServletRequest.class);
-            when(request.getHeader(CommonHttpHeader.AUTH_TOKEN.getHeaderKey())).thenReturn("aaaa-aaaa-aaaa-aaaa-aaaaaaaa");
-            when(request.getRequestURL()).thenReturn(new StringBuffer("http://ponies.com/12345/ponies"));
+        protected abstract boolean delegatable();
+    }
+
+    public static class WhenAuthenticatingDelegatableRequests extends TestParent {
+
+        @Override
+        protected boolean delegatable() {
+            return true;
         }
 
         @Test
-        public void shouldReturnFilterDirector() {
-            assertNotNull("When authenticating correctly, a non-null filter directors must be returned", authHandler.authenticate(request));
+        public void shouldPassNullOrBlankCredentials() {
+            when(request.getRequestURI()).thenReturn("/start/");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertTrue("Auth component must set X Authorization header when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-authorization").iterator().next().equalsIgnoreCase("Proxy"));
+            assertTrue("Auth component must set X Identity Status to Indeterminate when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-identity-status").iterator().next().equalsIgnoreCase("Indeterminate"));
+            assertEquals("Auth component must pass requests with null or blank credentials when in delegated mode", FilterAction.PASS, requestDirector.getFilterAction());
         }
 
         @Test
-        public void shouldValidateUserAuthToken() {
+        public void shouldPassNullOrBlankAuthToken() {
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertTrue("Auth component must set X Authorization header when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-authorization").iterator().next().equalsIgnoreCase("Proxy accountId"));
+            assertTrue("Auth component must set X Identity Status to Indeterminate when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-identity-status").iterator().next().equalsIgnoreCase("Indeterminate"));
+            assertEquals("Auth component must pass requests with null or blank auth token when in delegated mode", FilterAction.PASS, requestDirector.getFilterAction());
+        }
+
+        @Test
+        public void shouldPassNullOrBlankAccountId() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertTrue("Auth component must set X Authorization header when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-authorization").iterator().next().equalsIgnoreCase("Proxy"));
+            assertTrue("Auth component must set X Identity Status to Indeterminate when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-identity-status").iterator().next().equalsIgnoreCase("Indeterminate"));
+            assertEquals("Auth component must pass requests with null or blank account id when in delegated mode", FilterAction.PASS, requestDirector.getFilterAction());
+        }
+
+        @Test
+        public void shouldRejectInvalidCredentials() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
+            when(authServiceClient.validateToken(any(Account.class), anyString())).thenReturn(Boolean.FALSE);
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("Auth component must reject invalid credentials with a 401 when in delegated mode", HttpStatusCode.UNAUTHORIZED, requestDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldPassValidCredentials() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
             when(authServiceClient.validateToken(any(Account.class), anyString())).thenReturn(Boolean.TRUE);
             when(authServiceClient.getGroups(anyString())).thenReturn(groups);
-            
-            final FilterDirector actual = authHandler.authenticate(request);
-            
-            assertNotNull("When authenticating correctly, a non-null filter directors must be returned", actual);
-            assertEquals("When authenticating correctly, a valid token must pass", FilterAction.PASS, actual.getFilterAction());
 
-            verify(authServiceClient, times(1)).validateToken(any(Account.class), anyString());
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertTrue("When groups exist, handler must add X-PP-Groups header", requestDirector.requestHeaderManager().headersToAdd().get(PowerApiHeader.GROUPS.getHeaderKey().toLowerCase()).contains("group-id"));
+            assertTrue("Handler must add X-PP-User header", requestDirector.requestHeaderManager().headersToAdd().get(PowerApiHeader.USER.getHeaderKey().toLowerCase()).contains("accountId"));
+            assertTrue("Auth component must set X Authorization header when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-authorization").iterator().next().equalsIgnoreCase("Proxy accountId"));
+            assertTrue("Auth component must set X Identity Status to Confirmed when in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-identity-status").iterator().next().equalsIgnoreCase("Confirmed"));
+            assertEquals("Auth component must pass valid credentials when in delegated mode", FilterAction.PASS, requestDirector.getFilterAction());
         }
-        
+    }
+
+    public static class WhenAuthenticatingNonDelegatableRequests extends TestParent {
+        @Override
+        protected boolean delegatable() {
+            return false;
+        }
+
+        @Test
+        public void shouldNotReturnNullFilterDirectorOnResponseHandling() {
+            final FilterDirector director = handler.handleResponse(request, response);
+
+            assertNotNull("FilterDirector should not be null", director);
+        }
+
         @Test
         public void shouldHandleAuthenticationServiceFailures() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
             when(authServiceClient.validateToken(any(Account.class), anyString())).thenThrow(new RuntimeException());
-                    
-            assertEquals("When auth service fails, repose should return a 500", HttpStatusCode.INTERNAL_SERVER_ERROR, authHandler.authenticate(request).getResponseStatus());
 
-            verify(authServiceClient, times(1)).validateToken(any(Account.class), anyString());
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("When auth service fails, repose should return a 500", HttpStatusCode.INTERNAL_SERVER_ERROR, requestDirector.getResponseStatus());
         }
-        
+
         @Test
-        public void shouldRetrieveGroupsWhenTokenIsValid() {
+        public void shouldRejectNullOrBlankCredentials() {
+            when(request.getRequestURI()).thenReturn("/start/");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("Auth component must reject requests with null or blank credentials when not in delegated mode", HttpStatusCode.UNAUTHORIZED, requestDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldRejectNullOrBlankAuthToken() {
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("Auth component must reject requests with null or blank auth token when not in delegated mode", HttpStatusCode.UNAUTHORIZED, requestDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldRejectNullOrBlankAccountId() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/");
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("Auth component must reject requests with null or blank account id when not in delegated mode", HttpStatusCode.UNAUTHORIZED, requestDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldRejectInvalidCredentials() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
+            when(authServiceClient.validateToken(any(Account.class), anyString())).thenReturn(Boolean.FALSE);
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertEquals("Auth component must reject invalid credentials with a 401 when in not in delegated mode", HttpStatusCode.UNAUTHORIZED, requestDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldPassValidCredentials() {
+            when(request.getHeader(anyString())).thenReturn("some-random-auth-token");
+            when(request.getRequestURI()).thenReturn("/start/accountId/resource");
             when(authServiceClient.validateToken(any(Account.class), anyString())).thenReturn(Boolean.TRUE);
             when(authServiceClient.getGroups(anyString())).thenReturn(groups);
-            
-//                        filterDirector.requestHeaderManager().putHeader(PowerApiHeader.GROUPS.getHeaderKey(), groups);
-//            filterDirector.requestHeaderManager().putHeader(PowerApiHeader.USER.getHeaderKey(), accountUsername);
-            final FilterDirector actual = authHandler.authenticate(request);
-            final Map<String, Set<String>> headersAdded = actual.requestHeaderManager().headersToAdd();
-            
-            assertTrue("When groups exist, handler must add X-PP-Groups header", headersAdded.get(PowerApiHeader.GROUPS.getHeaderKey().toLowerCase()).contains("group-id"));
-            assertTrue("When groups exist, handler must add X-PP-Groups header", headersAdded.get(PowerApiHeader.USER.getHeaderKey().toLowerCase()).contains("12345"));
+
+            final FilterDirector requestDirector = handler.handleRequest(request, response);
+
+            assertTrue("When groups exist, handler must add X-PP-Groups header", requestDirector.requestHeaderManager().headersToAdd().get(PowerApiHeader.GROUPS.getHeaderKey().toLowerCase()).contains("group-id"));
+            assertTrue("Handler must add X-PP-User header", requestDirector.requestHeaderManager().headersToAdd().get(PowerApiHeader.USER.getHeaderKey().toLowerCase()).contains("accountId"));
+            assertTrue("Auth component must set X Authorization header when not in delegated mode", requestDirector.requestHeaderManager().headersToAdd().get("x-authorization").iterator().next().equalsIgnoreCase("Proxy accountId"));
+            assertEquals("Auth component must pass valid credentials when not in delegated mode", FilterAction.PASS, requestDirector.getFilterAction());
+        }
+    }
+
+    public static class WhenHandlingResponseFromServiceInDelegatedMode extends TestParent {
+        @Override
+        protected boolean delegatable() {
+            return true;
+        }
+
+        @Test
+        public void shouldModifyDelegatedWwwAuthenticateHeaderOn401() {
+            when(response.getHeader(CommonHttpHeader.WWW_AUTHENTICATE.getHeaderKey())).thenReturn("Delegated");
+            when(response.getStatus()).thenReturn(401);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+            final String expected = "RackAuth Realm=\"API Realm\"";
+
+            assertEquals("Auth component must return a 401 when delegatable origin service returns a 401", HttpStatusCode.UNAUTHORIZED, responseDirector.getResponseStatus());
+            assertEquals("Auth component must modify WWW-Authenticate header when delegatable origin service returns a 401", expected, responseDirector.responseHeaderManager().headersToAdd().get(CommonHttpHeader.WWW_AUTHENTICATE.toString()).iterator().next());
+        }
+
+        @Test
+        public void shouldModifyDelegatedWwwAuthenticateHeaderOn403() {
+            when(response.getHeader(CommonHttpHeader.WWW_AUTHENTICATE.getHeaderKey())).thenReturn("Delegated");
+            when(response.getStatus()).thenReturn(403);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+            final String expected = "RackAuth Realm=\"API Realm\"";
+
+            assertEquals("Auth component must return a 403 when delegatable origin service returns a 403", HttpStatusCode.FORBIDDEN, responseDirector.getResponseStatus());
+            assertEquals("Auth component must modify WWW-Authenticate header when delegatable origin service returns a 403", expected, responseDirector.responseHeaderManager().headersToAdd().get(CommonHttpHeader.WWW_AUTHENTICATE.toString()).iterator().next());
+        }
+
+        @Test
+        public void shouldReturn500OnAuthFailureWith501() {
+            when(response.getHeader(CommonHttpHeader.WWW_AUTHENTICATE.getHeaderKey())).thenReturn("Delegated");
+            when(response.getStatus()).thenReturn(501);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+
+            assertEquals("Auth component must return a 500 when delegatable origin service returns a 501", HttpStatusCode.INTERNAL_SERVER_ERROR, responseDirector.getResponseStatus());
+        }
+    }
+
+    public static class WhenHandlingResponseFromServiceNotInDelegatedMode extends TestParent {
+        @Override
+        protected boolean delegatable() {
+            return false;
+        }
+
+        @Test
+        public void shouldReturn500OnAuthFailureWith401() {
+            when(response.getHeader(CommonHttpHeader.WWW_AUTHENTICATE.getHeaderKey())).thenReturn("Not-Delegate");
+            when(response.getStatus()).thenReturn(401);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+
+            assertEquals("Auth component must return a 500 when un-delegatable origin service returns a 401", HttpStatusCode.INTERNAL_SERVER_ERROR, responseDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldReturn500OnAuthFailureWith403() {
+            when(response.getStatus()).thenReturn(403);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+
+            assertEquals("Auth component must return a 500 when un-delegatable origin service returns a 403", HttpStatusCode.INTERNAL_SERVER_ERROR, responseDirector.getResponseStatus());
+        }
+
+        @Test
+        public void shouldReturn501OnAuthFailureWith501() {
+            when(response.getHeader(CommonHttpHeader.WWW_AUTHENTICATE.getHeaderKey())).thenReturn("Not-Delegate");
+            when(response.getStatus()).thenReturn(501);
+
+            final FilterDirector responseDirector = handler.handleResponse(request, response);
+
+            assertEquals("Auth component must return a 501 when un-delegatable origin service returns a 501", HttpStatusCode.NOT_IMPLEMENTED, responseDirector.getResponseStatus());
         }
     }
 }
